remove polyfills
