+++ package.json
@@ -4,7 +4,6 @@
   "version": "3.0.0",
   "devDependencies": {
     "@babel/core": "^7.18.13",
-    "@babel/polyfill": "^7.12.1",
     "@babel/preset-env": "^7.18.10",
     "@babel/preset-react": "^7.18.6",
     "@babel/preset-typescript": "^7.18.6",
@@ -43,11 +42,8 @@
     "lint-staged": "^13.0.3",
     "lodash": "^4.17.21",
     "mini-css-extract-plugin": "^2.6.1",
-    "mkdirp": "^1.0.4",
     "npm-run-all": "^4.1.5",
-    "object-assign": "^4.1.1",
     "prettier": "^2.7.1",
-    "raw-loader": "^4.0.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "sass": "^1.54.5",
@@ -58,8 +54,7 @@
     "uuid": "^9.0.0",
     "webpack": "^5.74.0",
     "webpack-cli": "^4.10.0",
-    "webpack-dev-server": "^4.10.0",
-    "whatwg-fetch": "^3.5.0"
+    "webpack-dev-server": "^4.10.0"
   },
   "scripts": {
     "webpack": "cross-env NODE_ENV=production webpack",
