+++ scripts/utils/fsAsync.js
@@ -1,13 +1,11 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: MIT-0
-const fs = require('fs');
-const path = require('path');
-const mkdirp = require('mkdirp');
-const { promisify } = require('util');
+const fs = require('node:fs/promises');
+const path = require('node:path');
 
 const writeFileAsync = async (...args) => {
-  await mkdirp(path.dirname(args[0]));
-  return promisify(fs.writeFile)(...args);
+  await fs.mkdir(path.dirname(args[0]), { recursive: true });
+  return fs.writeFile(...args);
 };
 
 module.exports = {
