+++ src/common/localStorage.ts
@@ -1,0 +1,15 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: MIT-0
+export const save = (key: string, value: any) => localStorage.setItem(key, JSON.stringify(value));
+
+export const load = (key: string) => {
+  const value = localStorage.getItem(key);
+  try {
+    return value && JSON.parse(value);
+  } catch (e) {
+    console.warn(
+      `⚠️ The ${key} value that is stored in localStorage is incorrect. Try to remove the value ${key} from localStorage and reload the page`
+    );
+    return undefined;
+  }
+};
