+++ src/pages/commons/use-local-storage.ts
@@ -1,0 +1,15 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: MIT-0
+import { useState } from 'react';
+import { load, save } from '../../common/localStorage';
+
+export function useLocalStorage<T>(key: string, defaultValue?: T) {
+  const [value, setValue] = useState(() => load(key) ?? defaultValue);
+
+  function handleValueChange(newValue: T) {
+    setValue(newValue);
+    save(key, newValue);
+  }
+
+  return [value, handleValueChange];
+}
