+++ src/pages/split-panel-comparison/index.jsx
@@ -14,7 +14,7 @@ import {
   TableEmptyState,
 } from '../commons/common-components';
 import { paginationLabels } from '../../common/labels';
-import { useLocalStorage } from '../../common/localStorage';
+import { useLocalStorage } from '../commons/use-local-storage';
 import { EC2ToolsContent, FullPageHeader } from '../table/common-components';
 import INSTANCES from '../../resources/ec2-instances';
 import { COLUMN_DEFINITIONS_MAIN, SELECTION_LABELS, DEFAULT_PREFERENCES, EC2Preferences } from './table-config';
